UiAutomationPython/requirements.txt
pytest
selenium
pyHamcrest
allure-pytest